New modules for argparse based operation, and changes to minimization methods.
This commit introduces the fitting.py and argument_parser.py modules to enable simpler
interaction with the fitting parts the prospector code, as outlined in the new
prospector_parse.py script. The main feature is an importable lnprobfn and several
fitting function wrappers.  The parser includes a number of common options.
In addition, the minimization algorithms have been simplified and both Pminimize and
pminimize removed. Finally, the prospector.py script has been modified to work with the
new minimization methods, and the demo_*params.py files now have the correct levenberg
spelling in run_params to trigger minimization.