Basic argument parsing operation.
This commit improves and renames the default argument parser, which now
includes options related to nested sampling and optimization.  It also
changes the logic in fit_model and enables emcee restarts with run_emcee.
The read_results methods can now use build_* methods in addition to
load_* methods.  The prospector_parse.py script is updated to work with
the new argparser and fit_model method.