Argparse integration and model_setup backwards compatibility.
This commit makes the prosect_args module package level. For
compatibility with script based operation it modifies model_setup
methods to be able to use build_x methods in parameter files,
but warns the user to switch to argparse style parameter files.
It also adds a method to read_results that will return the ingredients
necessary to restart an emcee run from a given results file.